syscallcompat: move asUser to its own file
